+++ src/max/mod.rs
@@ -6,6 +6,7 @@ pub use corpus::Corpus;
 pub use fuzzer::{Fuzzer, FuzzerConfig};
 pub use harness::MaxHarness;
 pub use sequence::Sequence;
+pub use shrinker::{Shrinker, ShrinkerConfig};
 pub use value::Value;
 
 pub mod best;
@@ -14,4 +15,5 @@ pub mod corpus;
 pub mod fuzzer;
 pub mod harness;
 pub mod sequence;
+pub mod shrinker;
 pub mod value;